requirements.txt
azure-functions
pandas>=1.3.0
numpy>=1.21.0